+++ RELEASENOTES.md
@@ -51,6 +51,9 @@
         either contain only a single control sequence, or have critical info
         split across multiple control sequences
         ([#2935](https://github.com/androidx/media/issues/2935)).
+    *   VobSub: Use `START` and `STOP` SPU control commands to derive cue
+        duration instead of always defaulting to 5s
+        ([#2935](https://github.com/androidx/media/issues/2935)).
 *   Metadata:
 *   Image:
 *   DataSource: